+++ lua/mason-registry/sources/github.lua
@@ -1,5 +1,4 @@
 local InstallLocation = require "mason-core.installer.InstallLocation"
-local Optional = require "mason-core.optional"
 local Result = require "mason-core.result"
 local _ = require "mason-core.functional"
 local fetch = require "mason-core.fetch"